requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
facebook-sdk==2.0.0
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
pkginfo==1.4.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.6.0
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.11.0
tqdm==4.19.6
twine==1.9.1
uritemplate==3.0.0
urllib3==1.22
